Fix bug #10273
Although this a problem that only manifests itself on windows the change is general
and it works anywhere.

The major change is to change the file redirection > to -o the specifies the output file.

At the same time it makes the call to lyx2lyx less cryptic, e.g. to revert to the 1.3
format we have:

\converter lyx lyx13x "python -tt $$s/lyx2lyx/lyx2lyx -t 221 $$i > $$o"     ""

now instead of
python -tt $$s/lyx2lyx/lyx2lyx -t 221 $$i > $$o

we use a call where the version to revert is explicit
python -tt $$s/lyx2lyx/lyx2lyx -V 1.3 -o $$o $$i

or we could write a longer, but more understandable form:
python -tt $$s/lyx2lyx/lyx2lyx --final_version 1.3 --output $$o $$i

FWIW I shuffled the order of the arguments just for the sake of readability,
to let $$i be the last argument.