Improve documentation of LFUN_BUFFER_ZOOM_{IN,OUT}
The default for ZOOM_OUT is -20 from a user perspective. That is,
the following are equivalent:

  buffer-zoom-out -20
  buffer-zoom-out

The reason for this is that the argument to ZOOM_OUT is treated the
exact same way as ZOOM_IN. The only way they differ is how the
default case is handled.

This commit also clarifies that (1) the argument may be positive or
negative and that (2) the argument is interpreted as percentage
points, not percent.