Fix bug #10273
The export of compressed files to previous lyx versions failed on windows.

The change in this case is to explicitly pass the output file since we know it.

At the same time take care of the fact that lyx2lyx knows what file format
versions corresponds to each lyx version.