Fix XHTML export of German Additional Features manual.
There was an oddity in the manual that exposed a problem with the
test for the "special case" of an inset all by itself in a pargraph.
If a font change is applied to that inset, we still need to open the
paragraph.

(cherry picked from commit 3be23d18b64d1409d14e0c1c95f4192e9a23e543)