Convert deprecated '\rm' to '\textrm' in examples
This change fixes the export of our FeynmanDiagrams.lyx example
files on the latest TeX Live 2016.

The IJMP(C|D).lyx examples compiled before the change but might as
well convert these instances of '\rm' also.

The changes were made with the help of the following command:

  find ./ -iregex .*lyx -exec sed -i 's/\\rm /\\textrm /g' {} \;

and the resulting diff was confirmed manually.

This commit is consistent with 3cecd4d3.