Fix compilation of fr/beamer.lyx with XeTeX
The compilation did not succeed before because of an é in an ERT
inset. This is changed to \'{e}.

Also correct spelling of "decouvert" to "découvert".

Alternatives would be to use a LyX note or a comment inset. Since
the original version uses ERT, I do not change that.

This commit fixes the following ctest:

  export/examples/fr/beamer-article_pdf4_texF

It likely also fixes the corresponding test for fr/beamer.lyx, but
since the default output format is set to pdf2 we do not run the
test.

For more information, see:
https://www.mail-archive.com/search?l=mid&q=20160919192413.ftnz4gk6dl5w3k5c%40Opti1604