Amend and extend 9f3518bc
The comment is now in a comment inset. This change is extended to
all versions of beamer.lyx.

Change tracking is not used because it only affects a comment
(consistent with aa81ae7e).

This change is made according to updated conversation at:
https://www.mail-archive.com/search?l=mid&q=nruubc%24uu0%241%40blaine.gmane.org

(cherry picked from commit 40c262cbd865805d35aba65e25e919fc8dee6e69)