Fix label display in linguistic examples
Ideally, we would use a proper counter here, but this seems to be
impossible with the current counter mechanism.

What we would need is a counter that is incremented when either of the
layouts is used. Currently, the counter seems to be stepped only inside
one layout.

The current fix is a candidate for stable.